feat:reservation fix code

+++ src/main/java/com/musat/musat_backend/repository/ReservationRepository.java
@@ -3,5 +3,5 @@ package com.musat.musat_backend.repository;
 import com.musat.musat_backend.entity.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
 }
